pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-loop"
version = "0.1.0"
description = "Agent Loop implementation for workspace-coder"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml",
    "akshare>=1.0",
    "quantchdb>=0.1.11",
    "pandas>=2.0",
    "numpy>=1.24",
    "torch>=2.0.0",
    "tqdm>=4.0",
    "openai>=1.0",
    "pyarrow>=14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.3",
    "tensorboard>=2.0",
    "ipython>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["agent"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
show_missing = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by ruff)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "UP008", # use |= for set union
]
